Update 02-set_config_filebeat.sh
Former-commit-id: 6a9d230accaebb97a7d1176b9f2ae8a50d4602d3

+++ wazuh/config/02-set_config_filebeat.sh
@@ -11,8 +11,10 @@ set -e
 WAZUH_FILEBEAT_MODULE=wazuh-filebeat-0.1.tar.gz
 
 echo "FILEBEAT - Copy Filebeat config file"
-cp filebeat.yml /etc/filebeat/filebeat.yml
-
+if ! [[ -L /etc/filebeat/filebeat.yml ]]
+    cp filebeat.yml /etc/filebeat/filebeat.yml
+else
+    echo "Not needed. File already exist."
 echo "FILEBEAT - Set permissions"
 
 chmod go-w /etc/filebeat/filebeat.yml
